+++ src/controller/paidController.ts
@@ -1,8 +1,7 @@
-import { NextFunction, Request, Response } from "express";
+import { Request, Response } from "express";
 import { Transaction } from "../models/transaction";
 import { sendPayment } from "../utils/blink";
 import { wallet } from "..";
-import { getEncodedToken } from "@cashu/cashu-ts";
 import { Claim } from "../models/claim";
 
 export async function paidController(
@@ -23,6 +22,9 @@ export async function paidController(
 ) {
   const { eventType, transaction } = req.body;
   if (eventType === "receive.lightning") {
+    const newAmount =
+      Number(transaction.settlementAmount) -
+      Math.floor(Math.max(Number(transaction.settlementAmount) / 100, 1));
     const reqHash = transaction.initiationVia.paymentHash;
     const internalTx = await Transaction.getTransactionByHash(reqHash);
     try {
@@ -31,13 +33,10 @@ export async function paidController(
       // TO-DO log failed payment and retry
     }
     const { proofs } = await wallet.requestTokens(
-      transaction.settlementAmount,
+      newAmount,
       internalTx.mint_hash,
     );
-    const encoded = getEncodedToken({
-      token: [{ mint: process.env.MINTURL!, proofs }],
-    });
-    Claim.createClaim(internalTx.user, encoded);
+    Claim.createClaim(internalTx.user, process.env.MINTURL!, proofs);
     res.sendStatus(200);
   } else {
     return res.sendStatus(200);
